+++ CybozuLabs-md5/CybozuLabs-md5-tests.ts
@@ -1,0 +1,9 @@
+/// <reference path="CybozuLabs-md5.d.ts" />
+
+var hash: string;
+hash = CybozuLabs.MD5.calc("abc");
+hash = CybozuLabs.MD5.calc("abc", CybozuLabs.MD5.BY_ASCII);
+hash = CybozuLabs.MD5.calc("abc", CybozuLabs.MD5.BY_UTF16);
+
+var version: string;
+version = CybozuLabs.MD5.VERSION;
